govern/requirements.txt
Django>=1.10.4
MySQL-python>=1.2.5

Pillow>=3.4.2

django-debug-toolbar>=1.6
django-import-export>=0.5.1

django-databrowse

django-grappelli>=2.9.1
django-filebrowser>=3.7.2

docutils>=0.13.1
gunicorn>=19.6.0

greenlet>=0.4.9
eventlet>=0.20.0
gevent>=1.1.2

xlwt
xlsxwriter

python-envcfg>=0.2.0

# django-reversion>=1.10.0

git+https://github.com/wujuguang/django-admin-honeypot.git
git+https://github.com/etianen/django-reversion.git
git+https://github.com/sshwsfc/django-xadmin.git
